refactor: middleware test moved from e2e to test_lit_server.
