+++ tests/e2e/test_e2e.py
@@ -80,14 +80,6 @@ def test_e2e_default_batching():
     assert resp.json() == {"output": 16.0}, "tests/simple_server.py didn't return expected output"
 
 
-@e2e_from_file("tests/e2e/default_middleware.py")
-def test_e2e_default_middleware():
-    resp = requests.post("http://127.0.0.1:8000/predict", json={"input": 4.0}, headers=None)
-    assert resp.status_code == 200, f"Expected response to be 200 but got {resp.status_code}"
-    assert resp.json() == {"output": 16.0}, "server didn't return expected output"
-    assert resp.headers["X-Request-Id"] == "00000"
-
-
 @e2e_from_file("tests/e2e/default_batched_streaming.py")
 def test_e2e_batched_streaming():
     resp = requests.post("http://127.0.0.1:8000/predict", json={"input": 4.0}, headers=None, stream=True)
